requirements.txt
langgraph>=0.2,<1.0
langchain-core>=0.3,<1.0
openai>=1.30,<2.0
anthropic>=0.30,<1.0
google-generativeai>=0.7,<1.0
streamlit>=1.35,<2.0
pydantic>=2.0,<3.0
pyyaml>=6.0
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.23
